Import (#50)
Implemented import for basic query shapes. This does not yet support negation as this is not yet implemented in the code base. There is a flag to indicate that not supported shapes can be safely ignored.

Removed support for python 3.10. It was too difficult to make mypy pass for both 3.10 and 3.11 because the torch versions are different. Some warnings coming up on 3.10 could be suppressed, but these suppressions are no longer needed on 3.11, leading to new warnings. 

---------

Co-authored-by: Michael Cochez <miselico@users.noreply.github.com>